Move bert benchmark stuff to a central place.
PiperOrigin-RevId: 269835917
