+++ official/benchmark/squad_evaluate_v1_1.py
@@ -1,6 +1,5 @@
 # Copyright 2019 Pranav Rajpurkar, Jian Zhang, Konstantin Lopyrev and
 # Percy Liang. All Rights Reserved.
-#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
